+++ src/Base/EasyField.php
@@ -3,11 +3,17 @@ namespace PlusTimeIT\EasyForms\Base;
 
 class EasyField
 {
+    protected $clearable = FALSE;
+
     protected $cols = 12;
 
     protected $component = 'v-text-field';
 
     protected $component_type = '';
+
+    protected $dense = TRUE;
+
+    protected $help = '';
 
     protected $label = '';
 
@@ -17,17 +23,19 @@ class EasyField
 
     protected $order = 0;
 
+    protected $outlined = TRUE;
+
     protected $placeholder = '';
 
     protected $readonly = FALSE;
 
     protected $required = TRUE;
 
+    protected $rules = [];
+
     protected $tooltip = '';
 
     protected $type = 'text';
-
-    protected $validation = [];
 
     protected $value = '';
 }