+++ src/Fields/TextField.php
@@ -13,7 +13,33 @@ class TextField extends EasyField implements FieldInterface
         return $this->setOptions($options);
     }
 
+    public function getCounter(): bool
+    {
+        return $this->counter;
+    }
+
+    public function getMaxLength(): int
+    {
+        return $this->maxlength;
+    }
+
+    public function setCounter(bool $counter): self
+    {
+        $this->counter = $counter;
+        return $this;
+    }
+
+    public function setMaxLength(int $maxlength): self
+    {
+        $this->maxlength = $maxlength;
+        return $this;
+    }
+
     protected $component = 'v-text-field';
+
+    protected $counter = FALSE;
+
+    protected $maxlength;
 
     use ConvertTrait;
     use FieldTrait;