+++ src/Traits/FieldTrait.php
@@ -3,6 +3,11 @@ namespace PlusTimeIT\EasyForms\Traits;
 
 trait FieldTrait
 {
+    public function getClearable(): bool
+    {
+        return $this->clearable;
+    }
+
     public function getCols(): int
     {
         return $this->cols;
@@ -16,6 +21,16 @@ trait FieldTrait
     public function getComponentType(): string
     {
         return $this->component_type;
+    }
+
+    public function getDense(): bool
+    {
+        return $this->dense;
+    }
+
+    public function getHelp(): string
+    {
+        return $this->help;
     }
 
     public function getLabel(): string
@@ -33,6 +48,11 @@ trait FieldTrait
         return $this->order;
     }
 
+    public function getOutlined(): bool
+    {
+        return $this->outlined;
+    }
+
     public function getPlaceholder(): string
     {
         return $this->placeholder;
@@ -48,6 +68,11 @@ trait FieldTrait
         return $this->required;
     }
 
+    public function getRules(): array
+    {
+        return $this->rules;
+    }
+
     public function getTooltip(): string
     {
         return $this->tooltip;
@@ -58,14 +83,15 @@ trait FieldTrait
         return $this->type;
     }
 
-    public function getValidation(): array
-    {
-        return $this->validation;
-    }
-
     public function getValue(): mixed
     {
         return $this->value;
+    }
+
+    public function setClearable(bool $clearable): self
+    {
+        $this->clearable = $clearable;
+        return $this;
     }
 
     public function setCols(int $cols): self
@@ -83,6 +109,18 @@ trait FieldTrait
     public function setComponentType(string $component_type): self
     {
         $this->component_type = $component_type;
+        return $this;
+    }
+
+    public function setDense(bool $dense): self
+    {
+        $this->dense = $dense;
+        return $this;
+    }
+
+    public function setHelp(string $help): self
+    {
+        $this->help = $help;
         return $this;
     }
 
@@ -104,6 +142,12 @@ trait FieldTrait
         return $this;
     }
 
+    public function setOutlined(bool $outlined): self
+    {
+        $this->outlined = $outlined;
+        return $this;
+    }
+
     public function setPlaceholder(string $placeholder): self
     {
         $this->placeholder = $placeholder;
@@ -122,6 +166,12 @@ trait FieldTrait
         return $this;
     }
 
+    public function setRules(array $rules): self
+    {
+        $this->rules = $rules;
+        return $this;
+    }
+
     public function setTooltip(string $tooltip): self
     {
         $this->tooltip = $tooltip;
@@ -131,12 +181,6 @@ trait FieldTrait
     public function setType(string $type): self
     {
         $this->type = $type;
-        return $this;
-    }
-
-    public function setValidation(array $validation): self
-    {
-        $this->validation = $validation;
         return $this;
     }